init fullrepo rewrite (WIP)

+++ utils/fullrepo.sh
@@ -3,36 +3,30 @@
 # doc: build a full copy of the pacman repo
 
 echo "building instantOS pacman repository"
-cd || exit 1
-
 # exit when a command fails
 set -eo pipefail
 
-sudo pacman -S --needed --noconfirm \
-    wmctrl \
-    xdotool \
-    go \
-    ninja \
-    meson \
-    check \
-    libnotify \
-    tk \
-    vala \
-    gobject-introspection \
-    vte3 \
-    dbus-glib \
-    appstream-glib \
-    archlinux-appstream-data-pamac \
-    libpamac-nosnap \
-    libindicator-gtk3 \
-    libindicator-gtk2
+cd
+
+# build functions
+source /usr/local/share/instanttools/utils.sh
+
+# install packages needed to build some AUR stuff
+installbuilddeps
 
 if [ -e instantbuild ]; then
-    echo "removing older build files"
-    rm -rf instantbuild
+    if imenu cli -c 'there are build files already present. Are you sure you want to remove them?'; then
+        echo "removing older build files"
+        rm -rf instantbuild
+    else
+        echo 'build cancelled'
+    fi
 fi
 
-mkdir ~/instantbuild
+CACHEDIR="$HOME/instantbuildcache/$(date '+%y/%m/%d')"
+
+mkdir -p "$CACHEDIR" || echo 'existing build cache found'
+mkdir ~/instantbuild || echo 'existing build directory found'
 
 # detect architecture
 UNAME="$(uname -m)"
@@ -40,25 +34,24 @@ if grep -q 'x8' <<<"$UNAME"; then
     echo "detected 64 bit build"
 elif grep -q '^i' <<<"$UNAME"; then
     echo "detected 32 bit build"
+    ARCH32='32'
+else
+    echo 'architecture is not supported, support should be easy to add though,
+feel free to send a PR to instanttools'
+    exit
 fi
 
-cd || exit
-mkdir stuff || echo "stuff existing" &>/dev/null
-cd stuff || exit
+mkdir ~/stuff || echo "stuff existing" &>/dev/null
+cd ~/stuff
 
 echo "removing old extra repo"
 [ -e extra ] && rm -rf extra
 
-cd ~/stuff || exit
 git clone --depth=1 https://github.com/instantos/extra.git
-cd ~/stuff/extra || exit
+cd extra
 rm -rf .git
 
 echo "starting instantOS repo build"
-# build functions
-source /usr/local/share/instanttools/utils.sh
-
-instantinstall paperbash || exit 1
 
 BUILDDIR="$(pwd)"
 
@@ -68,9 +61,9 @@ if [ -e aurpackages ]; then
         if grep -q ':' <<<"$i"; then
             AURNAME=$(echo $i | grep -o '^[^:]*')
             AURFINALNAME=$(echo $i | grep -o '[^:]*$')
-            aurbuild "$AURNAME" "$AURFINALNAME"
+            buildpackage "$AURNAME" "$AURFINALNAME"
         else
-            aurbuild "$i"
+            buildpackage "$i"
         fi
         cd "$BUILDDIR" || exit
     done
@@ -79,27 +72,11 @@ fi
 cd "$BUILDDIR" || exit
 for i in ./*; do
     if [ -e "$i/PKGBUILD" ]; then
-
-        if uname -m | grep -q '^i'; then
-            if [ -e "$i"/32ignore ]; then
-                echo "package $i is ignored on 32bit systems"
-                rm /tmp/pkgignore
-                continue
-            fi
-        fi
-
-        if [ -e "$i"/ignore ] || [ -e /tmp/pkgignore ]; then
-            echo "package $i is ignored"
-            rm /tmp/pkgignore
-            continue
-        fi
-
-        echo "building $i"
-        if ! bashbuild ${i#./}; then
-            echo "package $i build failed, exiting"
-        fi
+        buildpackage "$i"
     else
         echo "skipping folder $i, no PKGFILE found"
     fi
 done
+
+# TODO move built packages into instantbuild directory
